# src/users/dto/update-user.dto.ts
// TODO: use mapped types PartialType

import { PartialType } from '@nestjs/swagger';
import { CreateUserDto } from './create-user.dto';

export class UpdateUserDto extends PartialType(CreateUserDto) {}